requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
ruff
pytest
pytest-mock
pytest-cov
pytest-asyncio
httpx
pymupdf
sqlalchemy
neo4j
sqlite-vec
fastembed
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
langfuse
boto3
google-generativeai
ragas
pandas
datasets
nltk
arq
redis
locust
reportlab